feat(auth): add google-first sign-in page
Replace /auth with a minimal Google-first sign-in screen styled after the scene-engine UI.

Preserve the existing email/password screen at /auth/legacy and keep the Google CTA as a console-log placeholder for the upcoming OAuth flow.
