+++ frontend/src/features/auth/index.ts
@@ -2,6 +2,7 @@ export { AuthProvider, useAuth } from "./model/auth.context";
 export { default as RequireAuth } from "./routing/RequireAuth";
 export { default as RequireGuest } from "./routing/RequireGuest";
 export { default as AuthPage } from "./ui/AuthPage";
+export { default as LegacyAuthPage } from "./ui/LegacyAuthPage";
 export { AUTH_TABS, AUTH_ROUTES } from "./api/auth.constants";
 export {
   buildAuthRoute,
